configuration: Separate the legacy configuration
Move it into its own folder, includes are now prefixed with legacy/
