+++ softwareComponents/configuration/legacy/configuration/Configuration.cpp
