+++ softwareComponents/configuration/legacy/configuration/Generators.cpp
