+++ softwareComponents/configuration/legacy/configuration/Generators.h
@@ -1,4 +1,4 @@
-#include <Configuration.h>
+#include "Configuration.h"
 #include <optional>
 #include <unordered_set>
 #include <vector>
