+++ softwareComponents/kinematics/calculations.hpp
@@ -1,5 +1,5 @@
-#include <Configuration.h>
-#include <IO.h>
+#include <legacy/configuration/Configuration.h>
+#include <legacy/configuration/IO.h>
 
 #include <cassert>
 #include <cmath>
