+++ softwareComponents/snake_reconfig/Snake_algorithms.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <Configuration.h>
-#include <IO.h>
-#include <Generators.h>
+#include <legacy/configuration/Configuration.h>
+#include <legacy/configuration/IO.h>
+#include <legacy/configuration/Generators.h>
 #include <Algorithms.h>
 #include "MinMaxHeap.h"
 #include "Snake_structs.h"
