+++ softwareComponents/snake_reconfig/Snake_structs.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <Configuration.h>
-#include <Generators.h>
-#include <IO.h>
+#include <legacy/configuration/Configuration.h>
+#include <legacy/configuration/Generators.h>
+#include <legacy/configuration/IO.h>
 #include <cmath>
 #include <stack>
