+++ tools/emptyDockFill/emptyDockFill.cpp
@@ -1,5 +1,5 @@
-#include <Configuration.h>
-#include <IO.h>
+#include <legacy/configuration/Configuration.h>
+#include <legacy/configuration/IO.h>
 #include <iostream>
 #include <fstream>
 #include <string>
