+++ tools/tangle/main.cpp
@@ -1,6 +1,6 @@
-#include <Configuration.h>
-#include <Generators.h>
-#include <IO.h>
+#include <legacy/configuration/Configuration.h>
+#include <legacy/configuration/Generators.h>
+#include <legacy/configuration/IO.h>
 #include <dimcli/cli.h>
 #include <iostream>
 #include <random>
