+++ tools/visualizer/main.cpp
@@ -4,7 +4,7 @@
 #include "Visualizer.h"
 #include "Animator.h"
 #include <cxxopts.hpp>
-#include <IO.h>
+#include <legacy/configuration/IO.h>
 
 using namespace IO;
 //using Resolution = std::pair<int, int>;
